Add files via upload
This is the Raven Workspace where monitoR gunshot detections were evaluated and scored. The input data for this workspace is a folder labelled "Clips", which contains 9,263 6-second clips that contain possible gunshot detections that were output from monitoR. This folder of files is available upon request, as well as the Raven list file.